Prevent forceTranscodeV2 from returning false unless the media is in fact streamable.
This patch ensures/enforces the following relationship:
Format.isCompatible() = isSkipTranscode() && !forceTranscodeV2

The above relationship may be exploited in the future to enhance
the efficiency of the code. The override of isCompatible() by WEB.java
currently prevents this (it is the only "format" which breaks the
relationship).

Motivation for use as exploit:
See the addChild() method in DLNAResource.java and notice how the
conditional statements reduce when applying this relationship. The same
applies for FileTranscodeVirtualFolder.java.

Additionally, there is more control over when
RendererConfiguration.match() and Format.skip() are executed.