Tag resources with streamable subtitles
Resources need to be tagged so that other classes can be notified
of intentions for external subtitles to be streamed. This is necessary
to prevent subtitles from being both transcoded and streamed at the same
time.

Embedded subtitles don't need to be tagged. The decision to stream them
can be made within DLNAResource.java itself.