Improved subtitle control logic
Makes use of the newly defined method, class, and renderer configuration
option to control the decisions of whether video/subtitles should be
streamed and/or transcoded.

The general logic is to prioritise streaming of both video and
subtitles. If video is not supported, then both video and subtitles will
be transcoded.

For external subtitles, it is possible to over-ride the
transcoding of subtitles (i.e. stream subtitles along with transcoded
video) through the renderer configuration. Embedded subtitles do not
offer this feature.