Acquire DLNAMediaSubtitles earlier for ParserV1 renderers
Subtitles are parsed early on during a resources life-time. For some
reason, parserV1 will parse early but will not record the DLNAMediaInfo
for the subtitles until a much later stage in the resources life-time.
There does not seem to be a good reason for this?

These modifications address the issue by creating DLNAMediaInfo if
necessary and by storing the subtitles information.